refactor(cookbook): tidy lunch-concierge Dart layout and add IAP JWT gate
Split the Flutter client and shelf server by responsibility, add package
analysis_options, rename /healthz to /health for Cloud Run, and verify
IAP JWTs on /api/* when IAP_AUDIENCE is set (infra stays unchanged).
