+++ cookbook/lunch-concierge/client/pubspec.yaml
@@ -15,6 +15,8 @@ dependencies:
 
 dev_dependencies:
   flutter_lints: ^6.0.0
+  flutter_test:
+    sdk: flutter
 
 flutter:
   uses-material-design: true
