+++ cookbook/lunch-concierge/server/bin/server.dart
@@ -1,29 +1,28 @@
 import 'dart:io';
 
-import 'package:genkit_shelf/genkit_shelf.dart';
-import 'package:shelf/shelf.dart';
-import 'package:shelf/shelf_io.dart' as io;
-import 'package:shelf_router/shelf_router.dart';
-import 'package:shelf_static/shelf_static.dart';
-
+import 'package:lunch_concierge_server/app.dart';
 import 'package:lunch_concierge_server/lunch_flow.dart';
+import 'package:shelf/shelf_io.dart' as io;
 
 Future<void> main() async {
   final bundle = await createLunchFlow();
-  final router = Router()
-    ..get('/healthz', (_) => Response.ok('ok'))
-    ..post('/api/lunch', shelfHandler(bundle.flow));
-
-  final staticHandler = createStaticHandler(
-    Platform.environment['PUBLIC_DIR'] ?? '/app/public',
-    defaultDocument: 'index.html',
+  final publicDir = switch (Platform.environment['PUBLIC_DIR']) {
+    final value? when value.isNotEmpty => value,
+    _ => '/app/public',
+  };
+  final iapAudience = switch (Platform.environment['IAP_AUDIENCE']) {
+    final value? when value.isNotEmpty => value,
+    _ => null,
+  };
+  final handler = buildApp(
+    bundle: bundle,
+    publicDir: publicDir,
+    iapAudience: iapAudience,
   );
-
-  final handler = const Pipeline()
-      .addMiddleware(logRequests())
-      .addHandler(Cascade().add(router.call).add(staticHandler).handler);
-
-  final port = int.parse(Platform.environment['PORT'] ?? '8080');
+  final port = switch (Platform.environment['PORT']) {
+    final value? => int.parse(value),
+    _ => 8080,
+  };
   final server = await io.serve(handler, InternetAddress.anyIPv4, port);
   stdout.writeln('Lunch Concierge listening on :${server.port}');
 }
