+++ cookbook/lunch-concierge/server/pubspec.yaml
@@ -11,6 +11,8 @@ dependencies:
   genkit: ^0.14.1
   genkit_shelf: ^0.1.9
   genkit_vertexai: ^0.2.9
+  http: ^1.5.0
+  jose: ^0.3.5
   lunch_concierge_shared:
     path: ../shared
   postgres: ^3.5.12
@@ -20,3 +22,4 @@ dependencies:
 
 dev_dependencies:
   lints: ^6.0.0
+  test: ^1.26.0
